Fuzzy Selection

+++ db/migrate/20170718013014_create_course_settings.rb
@@ -1,7 +1,7 @@
 class CreateCourseSettings < ActiveRecord::Migration[5.1]
   def change
     create_table :course_settings do |t|
-      t.references :course_listing, type: :integer, foreign_key: true
+      t.references :course_listing, type: :bigint, foreign_key: true
       t.string :value_type
       t.integer :value_int
       t.string :value_string
